+++ f3-sidecar/main.go
@@ -1,0 +1,5 @@
+package main
+
+func main() {
+	println("F3 sidecar")
+}