+++ src/cli_shared/cli/client.rs
@@ -11,6 +11,8 @@ use directories::ProjectDirs;
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 
+use crate::daemon::db_util::ImportMode;
+
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(transparent)]
 #[cfg_attr(test, derive(derive_quickcheck_arbitrary::Arbitrary))]
@@ -37,18 +39,14 @@ impl Default for BufferSize {
 #[cfg_attr(test, derive(derive_quickcheck_arbitrary::Arbitrary))]
 pub struct Client {
     pub data_dir: PathBuf,
-    pub genesis_file: Option<String>,
+    pub genesis_file: Option<PathBuf>,
     pub enable_rpc: bool,
     pub enable_metrics_endpoint: bool,
     pub enable_health_check: bool,
-    /// If this is true, then we do not validate the imported snapshot.
-    /// Otherwise, we validate and compute the states.
-    pub snapshot: bool,
-    /// If this is true, delete the snapshot at `snapshot_path` if it's a local file.
-    pub consume_snapshot: bool,
     pub snapshot_height: Option<i64>,
     pub snapshot_head: Option<i64>,
     pub snapshot_path: Option<PathBuf>,
+    pub import_mode: ImportMode,
     /// Skips loading import CAR file and assumes it's already been loaded.
     /// Will use the CIDs in the header of the file to index the chain.
     pub skip_load: bool,
@@ -80,8 +78,7 @@ impl Default for Client {
             enable_metrics_endpoint: true,
             enable_health_check: true,
             snapshot_path: None,
-            snapshot: false,
-            consume_snapshot: false,
+            import_mode: ImportMode::default(),
             snapshot_height: None,
             snapshot_head: None,
             skip_load: false,