+++ src/libp2p/mod.rs
@@ -4,7 +4,7 @@
 mod behaviour;
 pub mod chain_exchange;
 mod config;
-mod discovery;
+pub mod discovery;
 mod gossip_params;
 pub mod hello;
 pub mod keypair;