+++ src/tool/subcommands/snapshot_cmd.rs
@@ -44,7 +44,24 @@ pub enum SnapshotCommands {
         vendor: snapshot::TrustedVendor,
     },
 
-    /// Validates the snapshot.
+    /// Validate the provided snapshots as a whole.
+    ValidateDiffs {
+        /// Number of recent epochs to scan for broken links
+        #[arg(long, default_value_t = 2000)]
+        check_links: u32,
+        /// Assert the snapshot belongs to this network. If left blank, the
+        /// network will be inferred before executing messages.
+        #[arg(long)]
+        check_network: Option<crate::networks::NetworkChain>,
+        /// Number of recent epochs to scan for bad messages/transactions
+        #[arg(long, default_value_t = 60)]
+        check_stateroots: u32,
+        /// Path to a snapshot CAR, which may be zstd compressed
+        #[arg(required = true)]
+        snapshot_files: Vec<PathBuf>,
+    },
+
+    /// Validate the snapshots individually.
     Validate {
         /// Number of recent epochs to scan for broken links
         #[arg(long, default_value_t = 2000)]
@@ -59,6 +76,9 @@ pub enum SnapshotCommands {
         /// Path to a snapshot CAR, which may be zstd compressed
         #[arg(required = true)]
         snapshot_files: Vec<PathBuf>,
+        /// Fail at the first invalid snapshot
+        #[arg(long)]
+        fail_fast: bool,
     },
 
     /// Make this snapshot suitable for use as a compressed car-backed blockstore.
@@ -124,7 +144,7 @@ impl SnapshotCommands {
                 }
                 Err(e) => cli_error_and_die(format!("Failed fetching the snapshot: {e}"), 1),
             },
-            Self::Validate {
+            Self::ValidateDiffs {
                 check_links,
                 check_network,
                 check_stateroots,
@@ -139,6 +159,42 @@ impl SnapshotCommands {
                     check_stateroots,
                 )
                 .await
+            }
+            Self::Validate {
+                check_links,
+                check_network,
+                check_stateroots,
+                snapshot_files,
+                fail_fast,
+            } => {
+                let mut has_fail = false;
+                for file in snapshot_files {
+                    println!("Validating {}", file.display());
+                    let result = async {
+                        let store = AnyCar::try_from(file.as_path())?;
+                        validate_with_blockstore(
+                            store.heaviest_tipset()?,
+                            Arc::new(store),
+                            check_links,
+                            check_network.clone(),
+                            check_stateroots,
+                        )
+                        .await?;
+                        Ok::<(), anyhow::Error>(())
+                    }
+                    .await;
+                    if let Err(e) = result {
+                        has_fail = true;
+                        eprintln!("Error: {:?}", e);
+                        if fail_fast {
+                            break;
+                        }
+                    }
+                }
+                if has_fail {
+                    bail!("validate failed");
+                };
+                Ok(())
             }
             Self::Compress {
                 source,